+++ src/simp.c
@@ -30,32 +30,32 @@ static char rcsid[] = "$Id$";
 	&&  r->op == CNST+TYPE && r->u.v.VAR == 0)\
 		return eqtree(OP, bittree(BAND, l->kids[0],\
 			cnsttree(unsignedtype, \
-				(unsigned long)fieldmask(l->u.field)<<fieldright(l->u.field))), r)
+				(unsigned long long)fieldmask(l->u.field)<<fieldright(l->u.field))), r)
 #define cfoldcnst(TYPE,VAR,OP) \
 	if (l->op == CNST+TYPE && r->op == CNST+TYPE) \
-		return cnsttree(inttype, (long)(l->u.v.VAR OP r->u.v.VAR))
+		return cnsttree(inttype, (long long)(l->u.v.VAR OP r->u.v.VAR))
 /* The address is folded as an integer and turned back into a pointer.  Doing
  * the arithmetic on the pointer itself is undefined when the base is null, and
  * a null base is exactly what `offsetof' asks a compiler to fold. */
 #define foldaddp(L,R,RTYPE,VAR) \
 	if (L->op == CNST+P && R->op == CNST+RTYPE) { \
 		Tree e = tree(CNST+P, ty, NULL, NULL);\
-		e->u.v.p = (void *)((unsigned long)L->u.v.p + (unsigned long)R->u.v.VAR);\
+		e->u.v.p = (void *)((unsigned long long)L->u.v.p + (unsigned long long)R->u.v.VAR);\
 		return e; }
 #define ufoldcnst(TYPE,EXP) if (l->op == CNST+TYPE) return EXP
 #define sfoldcnst(OP) \
 	if (l->op == CNST+U && r->op == CNST+I \
 	&& r->u.v.i >= 0 && r->u.v.i < 8*l->type->size) \
-		return cnsttree(ty, (unsigned long)(l->u.v.u OP r->u.v.i))
+		return cnsttree(ty, (unsigned long long)(l->u.v.u OP r->u.v.i))
 #define geu(L,R,V) \
 	if (R->op == CNST+U && R->u.v.u == 0) do { \
 		warning("result of unsigned comparison is constant\n"); \
-		return tree(RIGHT, inttype, root(L), cnsttree(inttype, (long)(V))); } while(0)
+		return tree(RIGHT, inttype, root(L), cnsttree(inttype, (long long)(V))); } while(0)
 #define idempotent(OP) if (l->op == OP) return l->kids[0]
 
 int needconst;
 int explicitCast;
-static int addi(long x, long y, long min, long max, int needconst) {
+static int addi(long long x, long long y, long long min, long long max, int needconst) {
 	int cond = x == 0 || y == 0
 	|| x < 0 && y < 0 && x >= min - y
 	|| x < 0 && y > 0
@@ -85,7 +85,7 @@ static int addd(double x, double y, double min, double max, int needconst) {
 
 }
 
-static Tree addrtree(Tree e, long n, Type ty) {
+static Tree addrtree(Tree e, long long n, Type ty) {
 	Symbol p = e->u.sym, q;
 
 	if (p->scope  == GLOBAL
@@ -124,7 +124,7 @@ static Tree addrtree(Tree e, long n, Type ty) {
 }
 
 /* div[id] - return 1 if min <= x/y <= max, 0 otherwise */
-static int divi(long x, long y, long min, long max, int needconst) {
+static int divi(long long x, long long y, long long min, long long max, int needconst) {
 	int cond = y != 0 && !(x == min && y == -1);
 	if (!cond && needconst) {
 		warning("overflow in constant expression\n");
@@ -150,7 +150,7 @@ static int divd(double x, double y, double min, double max, int needconst) {
 }
 
 /* mul[id] - return 1 if min <= x*y <= max, 0 otherwise */
-static int muli(long x, long y, long min, long max, int needconst) {
+static int muli(long long x, long long y, long long min, long long max, int needconst) {
 	int cond = x > -1 && x <= 1 || y > -1 && y <= 1
 	|| x < 0 && y < 0 && -x <= max/-y
 	|| x < 0 && y > 0 &&  x >= min/y
@@ -180,7 +180,7 @@ static int muld(double x, double y, double min, double max, int needconst) {
 
 }
 /* sub[id] - return 1 if min <= x-y <= max, 0 otherwise */
-static int subi(long x, long y, long min, long max, int needconst) {
+static int subi(long long x, long long y, long long min, long long max, int needconst) {
 	return addi(x, -y, min, max, needconst);
 }
 
@@ -225,18 +225,18 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 			identity(r,l,I,i,0);
 			break;
 		case CVI+I:
-			xcvtcnst(I,l->u.v.i,ty,i,(long)extend(l->u.v.i,ty));
+			xcvtcnst(I,l->u.v.i,ty,i,(long long)extend(l->u.v.i,ty));
 			break;
 		case CVU+I:
 			if (l->op == CNST+U) {
 				if (!explicitCast && l->u.v.u > ty->u.sym->u.limits.max.i)
 					warning("overflow in converting constant expression from `%t' to `%t'\n", l->type, ty);
 				if (needconst || !(l->u.v.u > ty->u.sym->u.limits.max.i))
-					return cnsttree(ty, (long)extend(l->u.v.u,ty));
+					return cnsttree(ty, (long long)extend(l->u.v.u,ty));
 			}
 			break;
 		case CVP+U:
-			xcvtcnst(P,(unsigned long)l->u.v.p,ty,u,(unsigned long)l->u.v.p);
+			xcvtcnst(P,(unsigned long long)l->u.v.p,ty,u,(unsigned long long)l->u.v.p);
 			break;
 		case CVU+P:
 			xcvtcnst(U,(void*)l->u.v.u,ty,p,(void*)l->u.v.u);
@@ -245,7 +245,7 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 			xcvtcnst(P,l->u.v.p,ty,p,l->u.v.p);
 			break;
 		case CVI+U:
-			xcvtcnst(I,l->u.v.i,ty,u,((unsigned long)l->u.v.i)&ones(8*ty->size));
+			xcvtcnst(I,l->u.v.i,ty,u,((unsigned long long)l->u.v.i)&ones(8*ty->size));
 			break;
 		case CVU+U:
 			xcvtcnst(U,l->u.v.u,ty,u,l->u.v.u&ones(8*ty->size));
@@ -257,7 +257,7 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 			xcvtcnst(U,l->u.v.u,ty,d,(long double)l->u.v.u);
 			break;
 		case CVF+I:
-			xcvtcnst(F,l->u.v.d,ty,i,(long)l->u.v.d);
+			xcvtcnst(F,l->u.v.d,ty,i,(long long)l->u.v.d);
 			break;
 		case CVF+F: {
 			float d;
@@ -283,13 +283,13 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 			commute(r,l);
 			identity(r,l,I,i,ones(8*ty->size));
 			if (r->op == CNST+I && r->u.v.u == 0)
-				return tree(RIGHT, ty, root(l), cnsttree(ty, 0L));
+				return tree(RIGHT, ty, root(l), cnsttree(ty, 0LL));
 			break;
 
 		case MUL+U:
 			commute(l,r);
 			if (l->op == CNST+U && (n = ispow2(l->u.v.u)) != 0)
-				return simplify(LSH, ty, r, cnsttree(inttype, (long)n));
+				return simplify(LSH, ty, r, cnsttree(inttype, (long long)n));
 			foldcnst(U,u,*);
 			identity(r,l,U,u,1);
 			break;
@@ -369,11 +369,11 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 			ufoldcnst(I,l->u.v.i ? cnsttree(ty, 1L) : cond(r));
 			break;
 		case BCOM+I:
-			ufoldcnst(I,cnsttree(ty, (long)extend((~l->u.v.i)&ones(8*ty->size), ty)));
+			ufoldcnst(I,cnsttree(ty, (long long)extend((~l->u.v.i)&ones(8*ty->size), ty)));
 			idempotent(BCOM+U);
 			break;
 		case BCOM+U:
-			ufoldcnst(U,cnsttree(ty, (unsigned long)((~l->u.v.u)&ones(8*ty->size))));
+			ufoldcnst(U,cnsttree(ty, (unsigned long long)((~l->u.v.u)&ones(8*ty->size))));
 			idempotent(BCOM+U);
 			break;
 		case BOR+U:
@@ -412,7 +412,7 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 			if (r->op == CNST+U && r->u.v.u == 0)
 				break;
 			if (r->op == CNST+U && (n = ispow2(r->u.v.u)) != 0)
-				return simplify(RSH, ty, l, cnsttree(inttype, (long)n));
+				return simplify(RSH, ty, l, cnsttree(inttype, (long long)n));
 			foldcnst(U,u,/);
 			break;
 		case EQ+F:
@@ -453,7 +453,7 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 			if (l->op == CNST+I && r->op == CNST+I
 			&& r->u.v.i >= 0 && r->u.v.i < 8*l->type->size
 			&& muli(l->u.v.i, 1<<r->u.v.i, ty->u.sym->u.limits.min.i, ty->u.sym->u.limits.max.i, needconst))
-				return cnsttree(ty, (long)(l->u.v.i<<r->u.v.i));
+				return cnsttree(ty, (long long)(l->u.v.i<<r->u.v.i));
 			if (r->op == CNST+I && (r->u.v.i >= 8*ty->size || r->u.v.i < 0)) {
 				warning("shifting an `%t' by %d bits is undefined\n", ty, r->u.v.i);
 				break;
@@ -485,7 +485,7 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 				break;
 			xfoldcnst(I,i,%,divi);
 			if (r->op == CNST+I && r->u.v.i == 1)	/* l%1 => (l,0) */
-				return tree(RIGHT, ty, root(l), cnsttree(ty, 0L));
+				return tree(RIGHT, ty, root(l), cnsttree(ty, 0LL));
 			break;
 		case MOD+U:		
 			if (r->op == CNST+U && ispow2(r->u.v.u)) /* l%2^n => l&(2^n-1) */
@@ -511,7 +511,7 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 					simplify(MUL, ty, l, r->kids[1]));
 			if (l->op == CNST+I && l->u.v.i > 0 && (n = ispow2(l->u.v.i)) != 0)
 				/* 2^n * r => r<<n */
-				return simplify(LSH, ty, r, cnsttree(inttype, (long)n));
+				return simplify(LSH, ty, r, cnsttree(inttype, (long long)n));
 			identity(r,l,I,i,1);
 			break;
 		case NE+F:
@@ -544,7 +544,7 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 			identity(r,l,I,i,0);
 			if (l->op == CNST+I && r->op == CNST+I
 			&& r->u.v.i >= 0 && r->u.v.i < 8*l->type->size) {
-				long n = l->u.v.i>>r->u.v.i;
+				long long n = l->u.v.i>>r->u.v.i;
 				if (l->u.v.i < 0)
 					n |= ~0UL<<(8*l->type->size - r->u.v.i);
 				return cnsttree(ty, n);
@@ -577,10 +577,10 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 			break;
 		case SUB+P:
 			if (l->op == CNST+P && r->op == CNST+P)
-				return cnsttree(ty, (long)((char *)l->u.v.p - (char *)r->u.v.p));
+				return cnsttree(ty, (long long)((char *)l->u.v.p - (char *)r->u.v.p));
 			if (r->op == CNST+I || r->op == CNST+U)
 				return simplify(ADD, ty, l,
-					cnsttree(inttype, r->op == CNST+I ? -r->u.v.i : -(long)r->u.v.u));
+					cnsttree(inttype, r->op == CNST+I ? -r->u.v.i : -(long long)r->u.v.u));
 			if (isaddrop(l->op) && r->op == ADD+I && r->kids[1]->op == CNST+I)
 				/* l - (x + c) => l-c - x */
 				return simplify(SUB, ty,
@@ -591,7 +591,7 @@ Tree simplify(int op, Type ty, Tree l, Tree r) {
 	return tree(op, ty, l, r);
 }
 /* ispow2 - if u > 1 && u == 2^n, return n, otherwise return 0 */
-int ispow2(unsigned long u) {
+int ispow2(unsigned long long u) {
 	int n;
 
 	if (u > 1 && (u&(u-1)) == 0)
